requirements.txt
Flask
Werkzeug
google-cloud-speech
google-cloud-texttospeech
google-cloud-language
